% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_gradient_iceberg.R
\name{scale_color_gradient_iceberg}
\alias{scale_color_gradient_iceberg}
\title{scale color gradient iceberg}
\usage{
scale_color_gradient_iceberg()
}
\value{
a ggplot object which scales gradients
}
\description{
A function used to make a custom scale_color_gradient() using iceberg_palette
}
