% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_count_proportion.R
\name{full_count_proportion}
\alias{full_count_proportion}
\title{full_count_proportion()}
\usage{
full_count_proportion(df, column_name, round = FALSE)
}
\arguments{
\item{df:}{a dataframe}

\item{column_name:}{the name of the numerical column to  calculate the total sum from all rows and the proportion}
}
\value{
df: data frame with two new columns
}
\description{
A short description...
}
\details{
A helper function
}
\examples{
library(iceberg_package)

df \%>\% full_count_proportion(., number, round = TRUE)

}
