% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col.R
\name{get_col}
\alias{get_col}
\title{get_col()}
\usage{
get_col(name)
}
\arguments{
\item{name}{name of the color whose hash you want to return}
}
\value{
a color's hash in the iceberg_palette
}
\description{
A helper function
}
\examples{
library(iceberg_package)

get_col(light_grey)

}
