% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceberg_theme.R
\name{iceberg_theme}
\alias{iceberg_theme}
\title{The Iceberg Theme}
\usage{
iceberg_theme()
}
\value{
a ggplot theme object
}
\description{
A theme developed in R which includes a color palette modeled after icebergs
}
\examples{
# library(iceburg_theme)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point() + 
  iceburg_theme()

}
