% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fiill_iceberg.R
\name{scale_fill_iceberg}
\alias{scale_fill_iceberg}
\title{scale_fill_iceberg}
\usage{
scale_fill_iceberg()
}
\value{
a ggplot object used to scale colors
}
\description{
A palette developed to be used in the iceberg_package()
}
